+++ CHANGELOG.md
@@ -5,14 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0] - 2018-06-22
 
 ### Added
-- **dashboard** Tooltip for 'live' project badge
+- **dashboard** tooltip for 'live' project badge
 - `ops sync` command
 - `ops system refresh-certs` command
 - `ops system refresh-config` command
 - `ops system refresh-services` command
+- `ops mariadb create` command
+- `ops psql create` command
+
+### Changed
+- `ops psql` is now `ops psql cli`
+- `ops psql-import` is now `ops psql import`
+- `ops psql-export` is now `ops psql export`
+- `ops mariadb` is now `ops mariadb cli`
+- `ops mariadb-import` is now `ops mariadb import`
+- `ops mariadb-export` is now `ops mariadb export`
+
+
 
 ## [0.5.0] - 2018-05-24